+++ backend/src/main/java/reviewme/review/service/exception/ReviewGroupNotFoundByReviewRequestCodeException.java
@@ -1,4 +1,4 @@
-package reviewme.review.domain.exception;
+package reviewme.review.service.exception;
 
 import lombok.extern.slf4j.Slf4j;
 import reviewme.global.exception.NotFoundException;
