+++ backend/src/main/java/reviewme/review/service/module/TextAnswerValidator.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import reviewme.question.repository.QuestionRepository;
 import reviewme.review.domain.TextAnswer;
-import reviewme.review.domain.exception.InvalidTextAnswerLengthException;
+import reviewme.review.service.exception.InvalidTextAnswerLengthException;
 import reviewme.review.service.exception.SubmittedQuestionNotFoundException;
 
 @Component
