+++ backend/src/main/java/reviewme/template/service/module/TemplateMapper.java
@@ -14,9 +14,9 @@ import reviewme.template.domain.Section;
 import reviewme.template.domain.SectionQuestion;
 import reviewme.template.domain.Template;
 import reviewme.template.domain.TemplateSection;
-import reviewme.template.domain.exception.MissingOptionItemsInOptionGroupException;
-import reviewme.template.domain.exception.SectionInTemplateNotFoundException;
-import reviewme.template.domain.exception.TemplateNotFoundByReviewGroupException;
+import reviewme.template.service.exception.MissingOptionItemsInOptionGroupException;
+import reviewme.template.service.exception.SectionInTemplateNotFoundException;
+import reviewme.template.service.exception.TemplateNotFoundByReviewGroupException;
 import reviewme.template.repository.SectionRepository;
 import reviewme.template.repository.TemplateRepository;
 import reviewme.template.service.dto.response.OptionGroupResponse;
