+++ backend/src/test/java/reviewme/api/TemplateApiTest.java
@@ -12,7 +12,7 @@ import org.mockito.BDDMockito;
 import org.springframework.restdocs.mockmvc.RestDocumentationResultHandler;
 import org.springframework.restdocs.payload.FieldDescriptor;
 import org.springframework.restdocs.request.ParameterDescriptor;
-import reviewme.review.domain.exception.ReviewGroupNotFoundByReviewRequestCodeException;
+import reviewme.review.service.exception.ReviewGroupNotFoundByReviewRequestCodeException;
 
 class TemplateApiTest extends ApiTest {
