+++ backend/src/test/java/reviewme/review/service/module/TextAnswerValidatorTest.java
@@ -11,7 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import reviewme.question.domain.Question;
 import reviewme.question.repository.QuestionRepository;
 import reviewme.review.domain.TextAnswer;
-import reviewme.review.domain.exception.InvalidTextAnswerLengthException;
+import reviewme.review.service.exception.InvalidTextAnswerLengthException;
 import reviewme.review.service.exception.SubmittedQuestionNotFoundException;
 import reviewme.support.ServiceTest;
