+++ backend/src/test/java/reviewme/template/service/TemplateMapperTest.java
@@ -23,8 +23,8 @@ import reviewme.reviewgroup.domain.ReviewGroup;
 import reviewme.reviewgroup.repository.ReviewGroupRepository;
 import reviewme.support.ServiceTest;
 import reviewme.template.domain.Section;
-import reviewme.template.domain.exception.MissingOptionItemsInOptionGroupException;
-import reviewme.template.domain.exception.SectionInTemplateNotFoundException;
+import reviewme.template.service.exception.MissingOptionItemsInOptionGroupException;
+import reviewme.template.service.exception.SectionInTemplateNotFoundException;
 import reviewme.template.repository.SectionRepository;
 import reviewme.template.repository.TemplateRepository;
 import reviewme.template.service.dto.response.QuestionResponse;
